DayTrading/requirements.txt
alpaca-py>=0.13.0
anthropic>=0.25.0
websocket-client>=1.7.0
matplotlib>=3.8.0
pytest>=8.0.0
